requirements.txt
google-cloud-bigquery==3.12.0
mysql-connector-python==8.1.0
pandas==2.0.3
sqlalchemy==2.0.23
pyarrow==13.0.0
google-cloud-secret-manager==2.16.2
python-dotenv==1.0.0